Revert "debug render bullet and shoot order"
This reverts commit 76a5420571b301cedfa82c8e3266a937c2a49217.

# Conflicts:
#	src/GameField.java
#	src/Tower.java
#	src/TowerDefense.java

+++ src/Point.java
@@ -1,6 +1,6 @@
 public class Point {
-    double x;
-    double y;
+    int x;
+    int y;
 
     public void setX(int x) {
         this.x = x;
@@ -15,7 +15,7 @@ public class Point {
         this.y = y;
     }
 
-    public static double distance(double x1, double y1, double x2, double y2) {
+    public static double distance(int x1, int y1, int x2, int y2) {
         return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2));
     }
